chore : persistence 세부 모듈 생성(#26)
