+++ Infrastructure-Module/persistence/src/main/java/com/bamyanggang/persistence/experience/jpa/entity/ExperienceContentJpaEntity.java
@@ -1,0 +1,36 @@
+package com.bamyanggang.persistence.experience.jpa.entity;
+
+import static jakarta.persistence.FetchType.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.util.UUID;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Table(name = "experience_content")
+public class ExperienceContentJpaEntity {
+    @Id
+    @Column(name = "experience_content_id")
+    private UUID id;
+
+    @Column(columnDefinition = "TEXT")
+    private String question;
+
+    @Column(columnDefinition = "TEXT")
+    private String answer;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "experience_id")
+    private ExperienceJpaEntity experience;
+}
