+++ Infrastructure-Module/persistence/src/main/java/com/bamyanggang/persistence/experience/jpa/entity/ExperienceJpaEntity.java
@@ -1,0 +1,35 @@
+package com.bamyanggang.persistence.experience.jpa.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import java.util.UUID;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Table(name = "experience")
+public class ExperienceJpaEntity{
+    @Id
+    @Column(name = "experience_id")
+    private UUID id;
+
+    private String title;
+
+    private LocalDateTime startedAt;
+
+    private LocalDateTime endedAt;
+
+    /*
+     * Todo: User 연관관계 구현 필요
+     */
+}
+
+
