+++ Infrastructure-Module/persistence/src/main/java/com/bamyanggang/persistence/user/jpa/repository/UserJpaRepository.java
@@ -1,0 +1,10 @@
+package com.bamyanggang.persistence.user.jpa.repository;
+
+import com.bamyanggang.persistence.user.jpa.entity.UserJpaEntity;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserJpaRepository extends JpaRepository<UserJpaEntity, UUID> {
+    UserJpaEntity findBySocialId(String socialId);
+
+}
