+++ build.gradle.kts
@@ -10,7 +10,6 @@ plugins {
     kotlin("kapt") version "1.9.23"
     kotlin("plugin.spring") version "1.9.23"
     kotlin("plugin.jpa") version "1.9.23"
-    kotlin("plugin.lombok") version "1.9.23"
 }
 
 val kotestVersion = "5.8.1"
@@ -34,7 +33,6 @@ subprojects {
     apply(plugin = "kotlin-spring")
     apply(plugin = "org.jetbrains.kotlin.plugin.spring")
     apply(plugin = "org.jetbrains.kotlin.plugin.jpa")
-    apply(plugin = "org.jetbrains.kotlin.plugin.lombok")
     apply(plugin = "org.springframework.boot")
     apply(plugin = "io.spring.dependency-management")
     apply(plugin = "java-test-fixtures")
